+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- *(Unreleased changes not logged at this time, use something akin to `git diff vX.Y.Z HEAD` using latest `vX.Y.Z`.)*
+- Renamed internal modules.
+- *(Unreleased changes might not logged at this point, use something akin to `git diff vX.Y.Z HEAD` using latest `vX.Y.Z`.)*
 
 
 ## [8.2.0] - 2026-04-17